adding jenkins k8s support

+++ jenkins/k8s/clean.sh
@@ -2,14 +2,17 @@
 
 #!/usr/bin/env bash
 
-# delete pods
-docker exec kubectl-support kubectl delete --all pods --namespace=jenkins
-# remove namespace
-RESPONSE=`echo `docker exec kubectl-support kubectl delete namespaces jenkins``
 
-if [ "${RESPONSE}" != "Error response from daemon: No such container: kubectl-support" ]; then
+CID=`docker container ls -aq -f "name=kubectl-support"`
+
+if [ ! -z "${CID}" ]; then
+# delete pods
+    docker exec kubectl-support kubectl delete --all pods --namespace=jenkins
+# remove namespace
+    docker exec kubectl-support kubectl delete namespaces jenkins
+
 # check status of the pods in every 30 seconds
-PODS_STATUS=echo `docker exec kubectl-support kubectl get pods --namespace=jenkins`
+    PODS_STATUS=echo `docker exec kubectl-support kubectl get pods --namespace=jenkins`
     while [ "${PODS_STATUS}" != "No resources found." ]
     do
         sleep 30
